dashboard_project/requirements_dashboard.txt
# Core Django y visualización
Django==3.2.24  # LTS version compatible con Python 3.9
plotly==5.18.0
django-plotly-dash==2.1.4

# Dependencias existentes (manteniendo versiones compatibles)
Flask==2.2.5  # Versión compatible con dash
Werkzeug==2.2.3  # Versión compatible con Flask 2.2.5
pandas==2.2.0
numpy==1.26.4
boto3==1.34.34
pyyaml==6.0.1

# Dependencias adicionales para el dashboard
dash==2.9.3  # Última versión compatible con Flask 2.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0